OEL-0000: Remove explicit version requirement for drupal/remote_stream_wrapper.
The package is needed for media_avportal.
The explicit version was needed to prevent the package to be upgraded to 1.6, where a patch from media_avportal would no longer apply.

Now a new version of media_avportal has been released, which requires drupal/remote_stream_wrapper:^2.0, and no longer attempts to patch it. Therefore the explicit version is no longer needed.