EWPP-4656: Temporarily revert AV Portal video test changes.
The title attribute position has changed in AV Portal video output and
the tests have been already updated. However, the changes are not yet
on the media_avportal branch used for D11 update. This revert is for
passing the tests temporarily and this commit should be removed after
the media_avportal D11 update branch is rebased or the new D11
compatible release of media_avportal is used.