EWPP-4656: Define loading attribute for remote video in gallery view mode.
In D10 the 'eager' loading attribute was set automatically if it was
not defined explicitly. This logic was removed in D11 and the default
value of the loading attribute setting 'lazy' is used. This produces
inconsistency between D10 and D11 tests (as oe_whitelabel tests for
eager).

Related: https://www.drupal.org/node/3275103